Move HammerPointStatesSubject and remove dictionary
